+++ lib/capybara_accessible_selectors/filters/described_by.rb
@@ -2,9 +2,11 @@
 
 Capybara::Selector::FilterSet[:capybara_accessible_selectors].instance_eval do
   node_filter(:described_by, valid_values: [String, Regexp]) do |node, value|
+    Capybara::Helpers.warn("described_by is deprecated in favour of accessible_description")
+
     next false unless node[:"aria-describedby"]
 
-    description = CapybaraAccessibleSelectors::Helpers.element_describedby(node)
+    description = node.accessible_description
     next true if value.is_a?(String) && description.include?(value)
     next true if value.is_a?(Regexp) && value.match?(description)