+++ lib/capybara_accessible_selectors/node/accessible_name.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "capybara_accessible_selectors/nokogiri/accessible_name"
+
 module CapybaraAccessibleSelectors
   module DriverNodeExtensions
     def accessible_name
@@ -15,11 +17,17 @@ module CapybaraAccessibleSelectors
 
   module SeleniumNodeExtensions
     def accessible_name
-      native.accessible_name&.strip || ""
+      # Safari is not returning a flat string
+      native.accessible_name.to_s.strip.gsub(/\s+/, " ")
+    rescue ::Selenium::WebDriver::Error::UnknownError
+      # Safari can throw this
+      ""
     end
   end
 
-  ::Capybara::Driver::Node.include DriverNodeExtensions
-  ::Capybara::Node::Element.include NodeElementExtensions
-  ::Capybara::Selenium::Node.include SeleniumNodeExtensions
+  module RackTestNodeExtensions
+    def accessible_name
+      Nokogiri::AccessibleName.resolve(native) || ""
+    end
+  end
 end