+++ spec/capybara/node/element/accessible_name_spec.rb
@@ -1,21 +1,22 @@
 # frozen_string_literal: true
 
 RSpec.describe Capybara::Node::Element, "#accessible_name" do
-  it "returns an accessible name" do
+  it "returns the accessible name of an element" do
     render <<~HTML
-      <article aria-labelledby="id2 id1" data-test-id="test">Contents</article>
-      <span id="id1">name</span>
-      <span id="id2">Accessible</span>
+      <element1 id="el1" role="button" aria-labelledby="el2">content</element1>
+      <element2 id="el2" hidden>
+        <element3 id="el3" hidden>hello</element3>
+      </element2>
     HTML
-
-    expect(find(:element, "article").accessible_name).to eq "Accessible name"
+    expect(find_by_id("el1").accessible_name).to eq "hello"
   end
 
-  it "returns an empty string if there no accessible name" do
+  it "returns no name as empty" do
     render <<~HTML
-      <div>foo</div>
+      <div>
+        contents
+      </div>
     HTML
-
     expect(find(:element, "div").accessible_name).to eq ""
   end
 end