+++ spec/selectors/combo_box_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-describe "combo_box selector" do
+describe "combo_box selector", skip_driver: :rack_test do
   before do
     visit "/combo_box.html"
   end
 
-  ["aria 1.0", "aria 1.1"].each do |label|
+  ["aria 1.0", "aria 1.1", "aria 1.2"].each do |label|
     context label do
       it "finds a combo box" do
         combo_box = find(:field, label)
@@ -271,7 +271,7 @@ describe "combo_box selector" do
 
         describe "currently_with" do
           it "fills in a combo box with an existing value" do
-            select_combo_box_option "Banana", from: label
+            select_combo_box_option "Banana", search: "Ban", from: label
             select_combo_box_option "Apple", from: label, currently_with: "Banana"
             expect do
               select_combo_box_option "Banana", from: label, currently_with: "Foo", wait: false
@@ -418,7 +418,7 @@ describe "combo_box selector" do
 
       describe "currently_with" do
         it "fills in a combo box with an existing value" do
-          select_combo_box_option "Banana", from: label
+          select_combo_box_option "Banana", search: "Ban", from: label
           select_combo_box_option "Apple", from: label, currently_with: "Banana"
           expect do
             select_combo_box_option "Banana", from: label, currently_with: "Foo", wait: false