+++ spec/selectors/modal_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe "modal selector" do
+describe "modal selector", skip_driver: :rack_test do
   before do
     visit "/dialog.html"
   end